Move puppeteer config stuff into its own file
Lots of our setup (user fixtures + debugging methods, etc) is
duplicated across files.

Created a new file to hold all of the shared code we use in our
puppeteer tests.
