+++ web/src/pages/__tests__/NoJS.test.js
@@ -1,40 +1,17 @@
-import { clickAndWait } from './puppeteer-utils'
+import {
+  baseUrl,
+  clickAndWait,
+  isDebugging,
+  user_en,
+  user_fr,
+} from './puppeteer-config'
 
 const puppeteer = require('puppeteer')
 
-const isDebugging = () => {
-  let debugging_mode = {
-    headless: false,
-    slowMo: 25,
-    devtools: false,
-  }
-
-  return process.env.NOJS && process.env.NOJS === 'debug'
-    ? debugging_mode
-    : { args: ['--no-sandbox'] }
-}
-
 let browser
 let page
-const baseUrl = 'http://localhost:3004'
 
 const fr = false
-
-const user_fr = {
-  fullName: 'Dominique Henri',
-  email: 'snc@exemple.com',
-  paperFileNumber: '123456',
-  explanation: "Voyage d'affaires!",
-}
-
-const user_en = {
-  fullName: 'Leanne Graham',
-  email: 'cdc@example.com',
-  paperFileNumber: '123456',
-  reason: 'workOrSchool',
-  explanation: 'On business trip!',
-  familyOption: 'Peter',
-}
 
 beforeAll(async () => {
   browser = await puppeteer.launch(isDebugging())
