--- web/src/pages/__tests__/puppeteer-utils.js
@@ -1,4 +1,0 @@
-export const clickAndWait = async (page, selector, duration = 100) => {
-  await page.click(selector)
-  await page.waitFor(duration)
-}
